backend/requirements.txt
# Mobeus Admin Dashboard Requirements


# Web framework
fastapi>=0.95.0
uvicorn>=0.21.1
jinja2>=3.1.2
python-multipart>=0.0.6

# Database
psycopg2-binary>=2.9.6
sqlalchemy>=2.0.9

# OpenAI API
openai>=1.0.0

# RAG components
chromadb>=0.4.6

# System monitoring
psutil>=5.9.5

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.3
requests>=2.28.2
websockets>=15.0.1
python-docx>=0.8.11
httpx>=0.24.1
tiktoken>=0.3.3

# Visualization
matplotlib>=3.7.1
pandas>=2.0.0